Use NATS for all RPC
This simplifies the architecture of the CI system.
