+++ examples/create-poller.json
@@ -1,0 +1,5 @@
+{
+    "remote": "https://github.com/run-ci/run.git",
+    "branch": "master",
+    "op": "create"
+}
