+++ examples/delete-poller.json
@@ -1,4 +1,5 @@
 {
     "remote": "https://github.com/run-ci/run.git",
-    "branch": "master"
+    "branch": "master",
+    "op": "delete"
 }
